+++ src/api/v1/routers/user.py
@@ -5,7 +5,7 @@ from src.api.v1.schemas.user import User, UserCreate, UserLogin
 from src.api.v1.schemas.token import Token
 from src.api.v1.core.database import get_session, Session
 from src.controllers.user import UserController
-from src.helpers.security import AuthManager
+from src.api.v1.core.security import AuthManager
 
 router = APIRouter()
