+++ src/api/v1/schemas/token.py
@@ -1,5 +1,6 @@
 
 from pydantic import BaseModel
+from datetime import datetime
 
 
 class Token(BaseModel):
